Rewrite user info and auth hook to use Redux selectors and actions
